CI: Test for Dual NIC BC with HA for phc2sys
Added a new test case for the DualNICBC in HA mode. The test case
implements this sequence:
- Turn down the slave interface in the primary BC.
- Wait for phc2sys logs that reflect the selection of the secondary
  BC's slave interface.
- Turn up the original slave interface in the primary BC.
- Wait for ph2sys logs that reflect the selection of the original slave
  interface of the primary BC.

In order for this test case to work, I've done some modifications in the
test framework:

- Added a function to create a separate third config for DualNIC Boundary
Clocks working in HA mode. This extra config has phc2sys config opts and
haProfiles in the ptpSettings field.
- Updated the discovery mechanism to detect the DualNICBC HA mode when
  two BC configs and the extra HA config for ph2sys are found.
- Refactored GetPodLogsRegex and created GetPodLogsRegexSince and moved
  the regex detection to a separate func findRegexInStream().